Handle numbers greater than 32-bit integers
OR-ing the number against `0` would OR it against a 32-bit integer, causing issues with numbers longer than 32-bits.
eg: `10000000000 | 0 == 1410065408`
`Math.trunc()` simply drops the decimal from a number, which appeared to be the intention of OR-ing with 0.

Also named the functions while I was there so stack traces can display a name.